+++ out/test/live.test.js
@@ -36,6 +36,7 @@
       content: 'welcome home',
       tags: [],
       position: 1,
+      positionNullable: null,
       category: 1,
       date: today
     },
@@ -45,6 +46,7 @@
       content: 'this is about jQuery',
       tags: ['jquery'],
       position: 2,
+      positionNullable: 2,
       category: 1,
       date: yesterday
     },
@@ -54,6 +56,7 @@
       content: 'this is about History.js',
       tags: ['jquery', 'html5', 'history'],
       position: 3,
+      positionNullable: 3,
       category: 1,
       date: tomorrow
     },
@@ -127,6 +130,21 @@
               return pass;
             }
           }).setSearchString('id:index').add(models);
+          actual = liveCollection.toJSON();
+          expected = [modelsObject.index];
+          return assert.deepEqual(actual, expected);
+        });
+        it('should support pill searches with null', function() {
+          var actual, expected, liveCollection;
+          liveCollection = queryEngine.createLiveCollection().setPill('positionNullable', {
+            prefixes: ['positionNullable:'],
+            callback: function(model, value) {
+              var pass, pillRegex;
+              pillRegex = queryEngine.createSafeRegex(value);
+              pass = pillRegex.test(model.get('positionNullable'));
+              return pass;
+            }
+          }).setSearchString('positionNullable:null').add(models);
           actual = liveCollection.toJSON();
           expected = [modelsObject.index];
           return assert.deepEqual(actual, expected);